randr: move it under Xext/
