+++ Xext/meson.build
@@ -60,3 +60,4 @@ subdir('composite')
 if build_xwin or build_xquartz
     subdir('pseudoramiX')
 endif
+subdir('randr')
