+++ Xext/randr/randrstr_priv.h
