+++ Xext/randr/rrcrtc.c
@@ -27,10 +27,10 @@
 #include "dix/dix_priv.h"
 #include "dix/request_priv.h"
 #include "dix/rpcbuf_priv.h"
-#include "randr/randrstr_priv.h"
-#include "randr/rrdispatch_priv.h"
 #include "os/bug_priv.h"
 #include "os/osdep.h"
+#include "Xext/randr/randrstr_priv.h"
+#include "Xext/randr/rrdispatch_priv.h"
 
 #include "swaprep.h"
 #include "mipointer.h"
