+++ Xext/randr/rrproperty.c
@@ -25,9 +25,9 @@
 
 #include "dix/dix_priv.h"
 #include "dix/request_priv.h"
-#include "randr/rrdispatch_priv.h"
+#include "Xext/randr/rrdispatch_priv.h"
+#include "Xext/randr/randrstr_priv.h"
 
-#include "randrstr_priv.h"
 #include "propertyst.h"
 #include "swaprep.h"
