+++ Xext/randr/rrprovider.c
@@ -28,8 +28,9 @@
 
 #include "dix/dix_priv.h"
 #include "dix/request_priv.h"
-#include "randr/randrstr_priv.h"
-#include "randr/rrdispatch_priv.h"
+
+#include "Xext/randr/randrstr_priv.h"
+#include "Xext/randr/rrdispatch_priv.h"
 
 #include "swaprep.h"
