+++ Xext/randr/rrscreen.c
@@ -24,8 +24,8 @@
 #include "dix/dix_priv.h"
 #include "dix/request_priv.h"
 #include "dix/server_priv.h"
-#include "randr/randrstr_priv.h"
-#include "randr/rrdispatch_priv.h"
+#include "Xext/randr/randrstr_priv.h"
+#include "Xext/randr/rrdispatch_priv.h"
 
 static CARD16
  RR10CurrentSizeID(ScreenPtr pScreen);
