+++ Xext/randr/rrtransform.c
@@ -22,7 +22,7 @@
 #include <dix-config.h>
 
 #include "include/rrtransform.h"
-#include "randr/randrstr_priv.h"
+#include "Xext/randr/randrstr_priv.h"
 
 void
 RRTransformInit(RRTransformPtr transform)
