+++ hw/xfree86/common/xf86platformBus.c
@@ -36,8 +36,8 @@
 
 #include "config/hotplug_priv.h"
 #include "dix/screenint_priv.h"
-#include "randr/randrstr_priv.h"
 #include "os/osdep.h"
+#include "Xext/randr/randrstr_priv.h"
 
 #include "os.h"
 #include "../os-support/linux/systemd-logind.h"
